+++ src/commonMain/kotlin/com/algolia/search/model/params/SearchParameters.kt
@@ -17,10 +17,4 @@ public interface SearchParameters : CommonSearchParameters, CommonParameters {
      * [Documentation][https://www.algolia.com/doc/api-reference/api-parameters/restrictSearchableAttributes/?language=kotlin]
      */
     public var restrictSearchableAttributes: List<Attribute>?
-
-    /**
-     * Relevancy score to apply to search in virtual index [0-100]. Bigger value means less, but more relevant results,
-     * lesser value - less relevant results.
-     */
-    public var relevancyStrictness: Int?
 }